bugfixes, updates

+++ src/build.js
@@ -21,45 +21,74 @@ function bundle(src, dest) {
     const bundleStream = createBundleStream(dest)
 
     browserify(src)
-      .transform('babelify', {
-        presets: ['@babel/preset-env'],
-      })
+      // TODO: Just give up on babel, which we may not even need?
+      // This creates globals that SES doesn't like
+      // .transform('babelify', {
+      //   presets: ['@babel/preset-env'],
+      // })
       .bundle((err, bundle) => {
 
         if (err) writeError(err)
 
-        // TODO: minification
+        // TODO: minification, probably?
         // const { error, code } = terser.minify(bundle.toString())
         // if (error) {
         //   writeError(error.message, error, dest)
         // }
-        // let strData = code.toString().trim()
+        // let strData = postProcess(code.toString())
 
-        let strData = bundle.toString().trim()
-        if (strData.length === 0) writeError(`Bundled code is empty.`, null, dest)
+        let strData = postProcess(bundle.toString())
 
-        bundleStream.end(strData, (err) => {
-          if (err) writeError(err.message, err, dest)
+        closeBundleStream(bundleStream, strData)
+        .then(() => {
           console.log(`Build success: '${src}' plugin bundled as '${dest}'`)
           resolve(true)
         })
+        .catch((err) => writeError(err.message, err, dest))
       })
   })
 }
 
+function postProcess (str) {
+  str = str.trim()
+  str = str.replace(/\.import\(/g, '["import"](')
+  // eval("require('pkg')") => require('pkg')
+  str = str.replace(/([^\w]+)eval\(["'`]{1}(require\([^)]*\))["'`]{1}\)/g, '$1$2')
+  // eval(stuff) => stuff
+  str = str.replace(/([^\w]+)eval(\([^)]*\))/g, '$1$2')
+  if (str.length === 0) throw new Error(`Bundled code is empty after postprocessing.`)
+  return str
+}
+
+async function closeBundleStream (stream, str) {
+
+  if (str.endsWith(';')) str = str.slice(0, -1)
+  if (str.startsWith('(') && str.endsWith(')')) {
+    str = '() => ' + str
+  } else {
+    str = '() => (\n' + str + '\n)'
+  }
+
+  stream.end(str, (err) => {
+    if (err) throw err
+  })
+}
+
 function createBundleStream (dest) {
-  const bundleStream = fs.createWriteStream(dest, {
+  const stream = fs.createWriteStream(dest, {
     autoClose: false,
     encoding: 'utf8',
   })
-  bundleStream.on('error', err => {
+  stream.on('error', err => {
     writeError(err.message, err, dest)
   })
-  return bundleStream
+  return stream
 }
 
 function writeError(msg, err, destFilePath) {
   logError('Write error: ' + msg, err)
-  fs.unlinkSync(destFilePath)
+  try {
+    if (destFilePath) fs.unlinkSync(destFilePath)
+  } catch (_err) {}
   process.exit(1)
 }
